Moved BeforeThrow to com.llamalad7.mixinextras.point
